data_pipeline/requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
PyMuPDF==1.24.9
holidays==0.51
pyarrow
azure-identity
azure-storage-file-datalake